SDL related interface encapsulation
